Renamed translation helper
